Removed old books references and recreated new index.js, places,js in server/src and placesController in server/controllers
