+++ client/src/pages/Feed/Feed.js
@@ -1,9 +1,14 @@
 import React, { useState, useEffect, useRef, Component } from 'react'
 import { Link } from 'react-router-dom'
 
+//  EXS just some starter code for getting current GeoLocation
+
 navigator.geolocation.getCurrentPosition(function (position) {
-  console.log('Latitude is :', position.coords.latitude)
-  console.log('Longitude is :', position.coords.longitude)
+  const myLat = position.coords.longitude
+  const myLong = position.coords.longitude
+  //console.log('Latitude is :', position.coords.latitude)
+  //console.log('Longitude is :', position.coords.longitude)
+  console.log(myLat, myLong)
 })
 
 function Feed () {
