+++ server/models/user.js
@@ -4,17 +4,18 @@ const bcrypt = require('bcryptjs')
 mongoose.promise = Promise
 
 // Define userSchema
+// EXS 18th July 2020 - Create reference from user to userInfo
 const userSchema = new Schema({
   firstName: { type: String, unique: false },
   lastName: { type: String, unique: false },
   username: { type: String, unique: false, required: false },
   password: { type: String, unique: false, required: false },
-  places: [
+  userInfo: [
     {
       // Store ObjectIds in the array
       type: Schema.Types.ObjectId,
       // The ObjectIds will refer to the ids in the places model
-      ref: 'Place'
+      ref: 'UserInfo_id'
     }
   ]
 })
