+++ server/routes/index.js
@@ -1,17 +1,17 @@
-const path = require("path");
-const router = require("express").Router();
-const authRoutes = require("./auth");
-const apiRoutes = require("./api");
+const path = require('path')
+const router = require('express').Router()
+const authRoutes = require('./auth')
+const apiRoutes = require('./api/index')
 
 // Auth Routes
-router.use('/auth', authRoutes);
+router.use('/auth', authRoutes)
 
 // API Routes
-router.use("/api", apiRoutes);
+router.use('/api', apiRoutes)
 
 // If no routes are hit, send the React app
-router.use(function(req, res) {
-  res.sendFile(path.join(__dirname, "../client/build/index.html"));
-});
+router.use(function (req, res) {
+  res.sendFile(path.join(__dirname, '../client/build/index.html'))
+})
 
-module.exports = router;
+module.exports = router
